rename: index 파일명으로 변경
