+++ src/components/common/ExplanationIndex/css/index.module.css
