+++ src/components/routes/bookmark/BookmarkList/index.tsx
@@ -4,7 +4,7 @@ import { useQuery } from '@tanstack/react-query';
 import MonsterCardList from '../../../common/MonsterCardList/MonsterCardList';
 import RegistrationAct from '../../../common/RegistrationAct';
 import styles from './css/index.module.css';
-import ExplanationIndex from '../../../common/ExplanationIndex/ExplanationIndex';
+import ExplanationIndex from '../../../common/ExplanationIndex';
 import { useSession } from 'next-auth/react';
 import LoadingSpinner from '../../../common/LoadingSpinner/LoadingSpinner';
 import EmptyList from './EmptyList';
